internal/gocdk: rename biome files
